+++ src/main/java/com/troika/groupon/dao/GroupProjectDao.java
@@ -73,5 +73,5 @@ public interface GroupProjectDao {
 	 * @param createUser
 	 * @return
 	 */
-	public List<Map<String, Object>> findProjectByCreateUser(Integer createUser);
+	public List<Map<String, Object>> findProjectByCreateUser(Long createUser);
 }
